requirements.txt
# Runtime dependencies — installed in venv for production and dev
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
Pillow>=10.4.0
psutil>=6.0.0
PyYAML>=6.0.1
pydantic>=2.8.0
websockets>=13.0
